[#594] Octo seed files should be automatically loaded (e.g. a newer Octo build is used for the first time)
Background
----------
* Till now, any time a new Octo build is installed, the user is expected to manually load the octo seed files.
  This is not user-friendly and is best avoided.
  - octo-seed.sql
  - octo-seed.zwr

  This commit loads the seed files automatically.

Fix
---
* `src/auto_load_octo_seed_if_needed.c` : This file is invoked at `octo`/`rocto` startup if we determine that
  this is the first time we are using this octo environment using the current commit SHA (stored in a
  global variable node `^%ydboctoocto("seedfmt")`. If so, we invoke the function `auto_load_octo_seed()`.

* `src/auto_load_octo_seed.c` : This loads the `octo-seed.sql` file by using `parse_line()` to read the
  queries one by one and then `run_query()` to run them. This also loads the `octo-seed.zwr` file by using
  `$ydb_dist/mupip` to do the load of the global nodes.

* `src/run_query.c` : This is now reworked so `CREATE TABLE` and `CREATE FUNCTION` commands when run in
  auto load mode (new global variable `config->in_auto_load_octo_seed`) to not issue `ERR_CANNOT_CREATE_TABLE`
  or `ERR_CANNOT_CREATE_FUNCTION` errors due to pre-existing tables/functions. In this case, the create
  command does an implicit drop first.

* `README.md` and `doc/intro.rst` had instructions to do the manual seed data load. Those are now removed
  as they are no longer needed.

* `src/lexer.l` : The `YY_INPUT` macro is now simplified and the `get_input()` function that it relied on
  is now moved to its own file `src/get_input.c` (from the file `src/readline_get_more.c`) to be consistent
  (every function in its own C file) and now has a `do/while` loop that was previously there in the `YY_INPUT`
  macro. This move eliminated the need for a `YY_NULL` macro. These changes most likely also resulted in
  a warning that previously showed up only on the CentOS build of Octo to disappear. The changes are in
  - tools/ci/expected_warnings-centos.ref
  - tools/ci/expected_warnings-centos_release.ref

* `src/CMakeLists.txt` : When octo is built with `DISABLE_INSTALL=ON`, the octo seed files are now
  built in the same directory where `octo`/`rocto` binaries exist (i.e. in `${PROJECT_BINARY_DIR}/src`
  instead of `${PROJECT_BINARY_DIR}`). This in turn helps `auto_load_octo_seed.c` determine the path
  of the seed file from the path of the currently running executable.

  When octo is built with `DISABLE_INSTALL=OFF`, the octo seed files were already installed in the
  directory `$ydb_dist/plugin/octo` so there were no changes needed in this area.

* A pre-existing issue in the following files is now fixed. This was observed while basing the new
  `auto_load_octo_seed_if_needed.c` file on them. The issue was that the for loop was never executing
  twice like it was originally intended. If two processes are started at exactly the same time and
  come to the for loop at the same time, they will both each end up doing an auto upgrade unnecessarily.
  It is very unlikely in practice so not documenting this as a user-visible issue.
  - src/auto_upgrade_binary_definition_if_needed.c
  - src/auto_upgrade_plan_definition_if_needed.c

* A pre-existing potential issue with not erasing the input buffer after running queries through
  `parse_line()` is now eliminated in the following files using a new `ERASE_INPUT_BUFFER` macro (defined
  in `src/octo.h`) that makes sure all evidence of previously executed queries are completely erased before
  we move on to executing queries for this octo/rocto invocation. I am not sure if there is a real issue or
  not so not documenting it other than in this commit message.
  - src/auto_upgrade_binary_function_definition.c
  - src/auto_upgrade_binary_table_definition.c

* `src/config.h` : A few globals `config->dry_run` and `config->allow_schema_changes` have their type
  changed from `int` to `boolean_t` since they can only take on boolean values. A new boolean variable
  `config->in_auto_load_octo_seed` tracks whether we are in the auto load phase or not.

* `src/gbldefs.c` and `src/errors.c` : All global variables in Octo were previously laid out in `src/errors.c`,
  a non-intuitive location. This is now fixed by moving it to a new file `src/gbldefs.c`. This also necessitated
  a new `src/rocto_gbldefs.c` which is the same as `src/gbldefs.c` version except that it defines some more
  rocto-only global variables using the `IS_ROCTO` macro. Therefore `rocto_gbldefs.c` just includes `gbldefs.c`
  but before that defines the `IS_ROCTO` macro. Just like is done for `rocto_errors.c` vs `errors.c`.

* `src/octo_init.c` : A call to the new `auto_load_octo_seed_if_needed()` was added. Additionally, some error
  messages were made more helpful by including more context.

* All `ERR_LIBCALL_WITH_ARG` usages and some `ERR_SYSCALL` usages were examined and the function name had a
  `()` suffix added to clearly indicate it was a function call.

* `src/parser/select.y` : An error about `OCTOONEROWTABLE` table not found used to say that the octo seed
  file load might have been missed out. This is now removed as that is automatically done.

* `src/rocto_helper_waitpid.c` : An unnecessary initialization of `wstatus` variable to 0 was removed.

* `tests/test_helpers.bash.in`, `tools/ci/build.sh` and `tools/ci/docker-install.sh` : Code that loaded the octo
  seed file was removed as it is taken care of automatically going forward.

* `tools/ci/build.sh` : If installing from binary tarball (using `octoinstall.sh`), then the `octo-seed.sql`
  and `octo-seed.zwr` files were being copied from `tests/fixtures`. This is incorrect as `convert_catalog_sql.sh`
  and `convert_catalog_zwr.sh` does some conversions to these files during the `cmake`/`make` build process
  and it is these converted files that should have been used as the source. This is a pre-existing issue that
  showed up now because the seed files are no longer manually loaded in `tools/ci/build/sh`. This is now fixed
  to use the `octo-seed.sql` and `octo-seed.zwr` files from `src/` (which is where the converted files are
  placed now) instead of from `tests/fixtures`.

* `tools/ci/build.sh` : A seemingly unnecessary step of loading `names.sql` and `names.zwr` was removed from
  this script. Whichever bats subtest needs these would need to do this in the test related setup script
  invocation (in `test_helpers.bash.in`). Besides, loading `names.sql` here was causing a `REC2BIG` error.
  Not yet sure why that happened but I suspect it is because the install is not yet complete at this point.
  In any case, since the erroring step is now removed, did not worry about investigating this further.

  Below was the error message for the record (https://gitlab.com/nars1/YDBOcto/-/jobs/1000232357#L2184).

  ```
  [ERROR] ../src/run_query.c:518 2021-02-01 18:23:42 : ERR_YOTTADB : YottaDB error: 150373602,(SimpleAPI),%YDB-E-REC2BIG, Record size (12906) is greater than maximum (4080) for region: DEFAULT,%YDB-I-GVIS, 		Global variable: ^%ydboctoschema("PG_CATALOG.PG_TYPE","binary",0)
  ```

* `tests/test_index_statement.bats.in` : A dummy `octo` invocation was added to avoid lot of verbose output
  from polluting the reference file (due to auto loading of octo seed files).

+++ src/run_query.c
@@ -149,8 +149,8 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 	char		 pid_buffer[INT64_TO_STRING_MAX]; /* assume max pid is 64 bits even though it is a 4-byte quantity */
 	boolean_t	 release_query_lock;
 	SqlStatement	 stmt;
-	SqlStatementType is_create_function;
-	boolean_t	 wrapInTp;
+	boolean_t	 ok_to_drop, wrapInTp;
+	SqlStatementType result_type;
 
 	// Assign cursor prior to parsing to allow tracking and storage of literal parameters under the cursor local variable
 	YDB_STRING_TO_BUFFER(config->global_names.schema, &schema_global);
@@ -210,7 +210,8 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 		CLEANUP_QUERY_LOCK_AND_MEMORY_CHUNKS(query_lock, memory_chunks, &cursor_ydb_buff);
 		return 1;
 	}
-	if (config->dry_run || (no_data_STATEMENT == result->type)) {
+	result_type = result->type;
+	if (config->dry_run || (no_data_STATEMENT == result_type)) {
 		CLEANUP_QUERY_LOCK_AND_MEMORY_CHUNKS(query_lock, memory_chunks, &cursor_ydb_buff);
 		return (YDB_OK != status);
 	}
@@ -219,7 +220,7 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 
 	cursor_used = TRUE; /* By default, assume a cursor was used to execute the query */
 	release_query_lock = TRUE;
-	switch (result->type) {
+	switch (result_type) {
 	// This effectively means select_STATEMENT, but we have to assign ID's inside this function
 	// and so need to propagate them out
 	case table_alias_STATEMENT:
@@ -327,7 +328,7 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 		ci_param1.address = routine_name;
 		ci_param1.length = sizeof(routine_name);
 		/* Currently read-only queries are not wrapped in TP and read-write queries are wrapped in TP. */
-		switch (result->type) {
+		switch (result_type) {
 		case table_alias_STATEMENT:
 		case set_operation_STATEMENT:
 			/* Read-only query */
@@ -335,7 +336,7 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 			break;
 		default:
 			/* Read-write query */
-			assert(insert_STATEMENT == result->type);
+			assert(insert_STATEMENT == result_type);
 			wrapInTp = TRUE;
 			break;
 		}
@@ -418,20 +419,23 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 		/* Wait 10 seconds for the exclusive DDL change lock */
 		status = ydb_lock_incr_s(TIMEOUT_DDL_EXCLUSIVELOCK, &query_lock[0], 1, &query_lock[1]);
 		CLEANUP_AND_RETURN_IF_NOT_YDB_OK(status, memory_chunks, buffer, spcfc_buffer, null_query_lock, &cursor_ydb_buff);
-		/* Note: Last parameter is NULL above as we do not have any query lock to release
+		/* Note: "null_query_lock" used above as we do not have any query lock to release
 		 * at this point since query lock grab failed.
 		 */
 		/* First get a ydb_buffer_t of the table name into "table_name_buffer" */
-		if (drop_table_STATEMENT == result->type) {
+		if (drop_table_STATEMENT == result_type) {
 			tablename = result->v.drop_table->table_name->v.value->v.reference;
 			table = NULL;
+			ok_to_drop = TRUE;
 		} else {
 			UNPACK_SQL_STATEMENT(table, result, create_table);
 			UNPACK_SQL_STATEMENT(value, table->tableName, value);
 			tablename = value->v.reference;
+			/* If auto load of octo-seed.sql is in progress, a CREATE TABLE must do a DROP TABLE. */
+			ok_to_drop = config->in_auto_load_octo_seed;
 		}
 		YDB_STRING_TO_BUFFER(tablename, table_name_buffer);
-		if (drop_table_STATEMENT == result->type) {
+		if (ok_to_drop) {
 			/* DROP TABLE */
 			/* Check if OIDs were created for this table.
 			 * If so, delete those nodes from the catalog now that this table is going away.
@@ -459,17 +463,20 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 				/* YDB_ERROR_CHECK would already have been done inside "drop_schema_from_local_cache()" */
 				CLEANUP_AND_RETURN(memory_chunks, buffer, spcfc_buffer, query_lock, &cursor_ydb_buff);
 			}
-		} else {
+		}
+		if (create_table_STATEMENT == result_type) {
 			/* CREATE TABLE case. More processing needed. */
 			int	  text_table_defn_length;
 			SqlTable *sql_table;
 
-			/* Check if table already exists */
-			sql_table = find_table(tablename);
-			if (NULL != sql_table) {
-				/* Table already exists. Issue error. */
-				ERROR(ERR_CANNOT_CREATE_TABLE, tablename);
-				CLEANUP_AND_RETURN(memory_chunks, buffer, spcfc_buffer, query_lock, &cursor_ydb_buff);
+			if (!ok_to_drop) {
+				/* Check if table already exists */
+				sql_table = find_table(tablename);
+				if (NULL != sql_table) {
+					/* Table already exists. Issue error. */
+					ERROR(ERR_CANNOT_CREATE_TABLE, tablename);
+					CLEANUP_AND_RETURN(memory_chunks, buffer, spcfc_buffer, query_lock, &cursor_ydb_buff);
+				}
 			}
 			out = open_memstream(&buffer, &buffer_size);
 			assert(out);
@@ -544,12 +551,11 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 		// Wait 10 seconds for the exclusive DDL change lock
 		status = ydb_lock_incr_s(TIMEOUT_DDL_EXCLUSIVELOCK, &query_lock[0], 1, &query_lock[1]);
 		CLEANUP_AND_RETURN_IF_NOT_YDB_OK(status, memory_chunks, buffer, spcfc_buffer, null_query_lock, &cursor_ydb_buff);
-		/* Note: Last parameter is NULL above as we do not have any query lock to release
+		/* Note: "null_query_lock" parameter used above as we do not have any query lock to release
 		 * at this point since query lock grab failed.
 		 */
 		/* First, get a ydb_buffer_t of the function name into "function_name_buffer" */
-		is_create_function = (create_function_STATEMENT == result->type);
-		if (is_create_function) {
+		if (create_function_STATEMENT == result_type) {
 			UNPACK_SQL_STATEMENT(function, result, create_function);
 			UNPACK_SQL_STATEMENT(value, function->function_name, value);
 		} else {
@@ -569,7 +575,7 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 		function_hash_buffer = &function_name_buffers[2];
 		YDB_STRING_TO_BUFFER(function_hash, function_hash_buffer);
 		// Add function hash to parse tree for later addition to logical plan
-		if (is_create_function) {
+		if (create_function_STATEMENT == result_type) {
 			SQL_STATEMENT(function->function_hash, value_STATEMENT);
 			MALLOC_STATEMENT(function->function_hash, value, SqlValue);
 			UNPACK_SQL_STATEMENT(value, function->function_hash, value);
@@ -590,7 +596,7 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 		/* Check if function with computed hash exists already or not */
 		status = ydb_data_s(&octo_global, 3, &function_name_buffers[0], &ret_value);
 		CLEANUP_AND_RETURN_IF_NOT_YDB_OK(status, memory_chunks, buffer, spcfc_buffer, query_lock, &cursor_ydb_buff);
-		if (!is_create_function) {
+		if (drop_function_STATEMENT == result_type) {
 			/* DROP FUNCTION */
 			if (0 == ret_value) {
 				char fn_name[1024];
@@ -601,7 +607,30 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 				ERROR(ERR_CANNOT_DROP_FUNCTION, fn_name);
 				CLEANUP_AND_RETURN(memory_chunks, buffer, spcfc_buffer, query_lock, &cursor_ydb_buff);
 			}
-			/* Drop the function in question since DROP FUNCTION was explicitly requested */
+			ok_to_drop = TRUE;
+		} else {
+			/* CREATE FUNCTION */
+			/* If auto load of octo-seed.sql is in progress, a CREATE FUNCTION must do a DROP FUNCTION. */
+			ok_to_drop = config->in_auto_load_octo_seed;
+			if (!ok_to_drop) {
+				if (0 != ret_value) {
+					char fn_name[1024];
+
+					/* Function already exists. Issue error. */
+					get_function_name_and_parmtypes(fn_name, sizeof(fn_name), function_name,
+									function->parameter_type_list);
+					ERROR(ERR_CANNOT_CREATE_FUNCTION, fn_name);
+					CLEANUP_AND_RETURN(memory_chunks, buffer, spcfc_buffer, query_lock, &cursor_ydb_buff);
+				}
+			} else {
+				/* It is okay for a function that is being created (using "CREATE FUNCTION" in "octo-seed.sql")
+				 * to already exist or not. No errors should be issued in either case. Handle it accordingly.
+				 */
+				ok_to_drop = (0 != ret_value); /* DROP FUNCTION only if it exists */
+			}
+		}
+		if (ok_to_drop) {
+			/* DROP FUNCTION explicitly requested or implicitly assumed as part of auto load of "octo-seed.sql" */
 			status = delete_function_from_pg_proc(function_name_buffer, function_hash_buffer);
 			if (0 != status) {
 				CLEANUP_AND_RETURN(memory_chunks, buffer, spcfc_buffer, query_lock, &cursor_ydb_buff);
@@ -620,22 +649,11 @@ int run_query(callback_fnptr_t callback, void *parms, boolean_t send_row_descrip
 				// YDB_ERROR_CHECK would already have been done inside "drop_schema_from_local_cache()"
 				CLEANUP_AND_RETURN(memory_chunks, buffer, spcfc_buffer, query_lock, &cursor_ydb_buff);
 			}
-		} else {
-			/* CREATE FUNCTION */
-			if (0 != ret_value) {
-				char fn_name[1024];
-
-				/* Function already exists. Issue error. */
-				get_function_name_and_parmtypes(fn_name, sizeof(fn_name), function_name,
-								function->parameter_type_list);
-				ERROR(ERR_CANNOT_CREATE_FUNCTION, fn_name);
-				CLEANUP_AND_RETURN(memory_chunks, buffer, spcfc_buffer, query_lock, &cursor_ydb_buff);
-			}
 		}
-		if (is_create_function) {
+		if (create_function_STATEMENT == result_type) {
 			int text_function_defn_length;
 
-			// CREATE FUNCTION case. More processing needed.
+			/* CREATE FUNCTION */
 			out = open_memstream(&buffer, &buffer_size);
 			assert(out);
 			text_function_defn_length = emit_create_function(out, result);
